Fixed compatibility issues

+++ src/uniform_point_rng.cpp
@@ -6,7 +6,7 @@
 uniform_point_rng::uniform_point_rng(T left, T top, T right, T bottom)
 {
     assert(left <= right and bottom <= top);
-    gen = std::mt19937{static_cast<long unsigned int>(time(0))};
+    gen = std::mt19937{static_cast<unsigned int>(time(0))};
     dist = std::uniform_real_distribution<T>(0.0, 1.0);
     minValue[0] = left, maxValue[0] = right;
     minValue[1] = bottom, maxValue[1] = top;
